Split companies/factory and companies/fixtures.
