+++ src/conftest.py
@@ -5,6 +5,12 @@ pytest_plugins = [
     "a12n.fixtures",
     "users.factory",
     "users.fixtures",
-    "companies.factory",
-    "companies.fixtures",
+    "companies.company_factory",
+    "companies.department_factory",
+    "companies.employee_factory",
+    "companies.stock_factory",
+    "companies.company_fixtures",
+    "companies.department_fixtures",
+    "companies.employee_fixtures",
+    "companies.stock_fixtures",
 ]
